requirements.txt
openai>=1.0.0
google-generativeai>=0.7.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.0.0
streamlit>=1.30.0
pytest>=8.0.0